2nd commit added JWT implementation and added search Query by name but not yet error free

+++ routes/Articles/delete/deleteArticle.js
@@ -1,6 +1,7 @@
 const express = require(`express`)
 const mongoose = require(`mongoose`)
 const createError = require(`http-errors`)
+const jwt = require("jsonwebtoken")
 const router = express.Router()
 
 const Article = require(`../../../models/articleModel/articlesModel`)
@@ -9,7 +10,19 @@ router.delete(`/deletearticle/:articleId`, async (req, res, next) => {
 
     const queryId = req.params.articleId
 
+    let token
+
+    if (req.headers.authorization && req.headers.authorization.startsWith(`Bearer`)) {
+        token = req.headers.authorization.split(` `)[1]
+    }
+
+    if (!token) {
+        return next(createError(401, `Please Log In.`))
+    }
+
     try {
+
+        jwt.verify(token, process.env.JWT_SECRET_KEY)
 
         const currentQuery = await Article.findOneAndDelete({_id: queryId})
